Phase 5: dtype fix + CI infra + §G2 collection cleanup [2.0]
* anc-with-ai: fix(dtype): normalize int16 capture to float32 not float64 [phase5]

Mirror of the anc-backend fix (near-identical fork): audio_capture.py and
predict_noise_type.py cast int16 PCM to float64, escaping the I/O boundary and
violating the float32 [-1,1] contract. Surfaced by the Phase 5 dtype audit.


* ci: fix pyaudio build dep (portaudio19-dev); bump upload-artifact v3->v4 + codeql-action v2->v3; add security-events perm for SARIF upload [pre-existing infra fix]


* ci(test): apply §G2 collection cleanup; remove dead verify/test scripts [pre-existing debt]

Deletes tests/integration/verify_classifier_pipeline.py + tests/unit/test_audio_system.py (ModuleNotFoundError: database_schema); pytest.ini drops verify_*.py from python_files (verify_wav.py kept on disk, uncollected). Clears all 3 collection errors. Residual pre-existing failures (unrelated to the dtype change) logged in INTEGRATION_REPORT §anc-with-ai-deferred.


---------

+++ src/utils/audio_capture.py
@@ -186,8 +186,9 @@ class AudioCapture:
         audio_bytes = b''.join(self.frames)
         audio_int16 = np.frombuffer(audio_bytes, dtype=np.int16)
 
-        # Normalize to float [-1, 1]
-        audio_float = audio_int16.astype(np.float64) / 32768.0
+        # Normalize to float [-1, 1]. float32 (not float64) per the audio dtype
+        # contract — dividing a float32 array by a Python scalar stays float32.
+        audio_float = audio_int16.astype(np.float32) / 32768.0
 
         return audio_float
